+++ ocrd/ocrd/processor/base.py
@@ -130,6 +130,11 @@ class Processor(object):
         self.ocrd_tool = ocrd_tool
         self.version = version
         self.workspace = workspace
+        # FIXME HACK would be better to use pushd_popd(self.workspace.directory)
+        # but there is no way to do that in process here since it's an
+        # overridden method. chdir is almost always an anti-pattern.
+        if self.workspace:
+            os.chdir(self.workspace.directory)
         self.input_file_grp = input_file_grp
         self.output_file_grp = output_file_grp
         self.page_id = None if page_id == [] or page_id is None else page_id